requirements.txt
appdata==2.1.2
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
django-admin-sortable==2.3
django-appconf==1.0.5
django-ckeditor==6.4.1
django-colorful==1.3
django-embed-video==1.4.4
django-imagekit==4.1.0
django-ipware==4.0.2
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-taggit==3.0.0
Django==4.0.4
git-ignore==1.0.2
idna==3.3
lxml==4.8.0
pilkit==2.0
Pillow==9.1.1
pip==22.1.1
python-gettext==4.0
python-slugify==6.1.2
requests==2.27.1
setuptools==62.3.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
to-requirements.txt==1.1.8
urllib3==1.26.9
wheel==0.37.1
